Fix win64 media_unittests build breakage in media/audio/win
It looks like fix r181108 for r180936 win64 chrome build break didn't fix break cascaded into media_unittests.

R=henrika@chromium.org
R=scherkus@chromium.org
R=tommi@chromium.org
BUG=175190
TEST=win64 media_unittest builds, linux media_unittests builds and passes


Review URL: https://chromiumcodereview.appspot.com/12221091

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@181582 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 7ba066c1f553c901b6cff2cabe539fcc906cee3e

+++ media/audio/win/core_audio_util_win_unittest.cc
@@ -288,7 +288,7 @@ TEST_F(CoreAudioUtilWinTest, SharedModeInitialize) {
                                                               &format)));
 
   // Perform a shared-mode initialization without event-driven buffer handling.
-  size_t endpoint_buffer_size = 0;
+  uint32 endpoint_buffer_size = 0;
   HRESULT hr = CoreAudioUtil::SharedModeInitialize(client, &format, NULL,
                                                    &endpoint_buffer_size);
   EXPECT_TRUE(SUCCEEDED(hr));
@@ -345,7 +345,7 @@ TEST_F(CoreAudioUtilWinTest, CreateRenderAndCaptureClients) {
   EDataFlow data[] = {eRender, eCapture};
 
   WAVEFORMATPCMEX format;
-  size_t endpoint_buffer_size = 0;
+  uint32 endpoint_buffer_size = 0;
 
   for (int i = 0; i < arraysize(data); ++i) {
     ScopedComPtr<IAudioClient> client;
@@ -394,7 +394,7 @@ TEST_F(CoreAudioUtilWinTest, FillRenderEndpointBufferWithSilence) {
   EXPECT_TRUE(client);
 
   WAVEFORMATPCMEX format;
-  size_t endpoint_buffer_size = 0;
+  uint32 endpoint_buffer_size = 0;
   EXPECT_TRUE(SUCCEEDED(CoreAudioUtil::GetSharedModeMixFormat(client,
                                                               &format)));
   CoreAudioUtil::SharedModeInitialize(client, &format, NULL,
